controllers->dataTables->various->getColumns: hid various columns on load. controllers->dataTables->various->getColumns: added titles. controllers->dataTables->various->getColumns: modified to have related name instead of magic number. controllers->dataTables->various->query: added related tables as required. controllers->dataTables->various->getColumns: modified response to link or image as appropriate. controllers->dataTables->various: Moved orkID to end. various: updated 'personas' to 'personae'. various: updated 'capitol' to 'capital'. controllers->api->various: added gate use. controllers->api->various->various: added gateway restrictions, as appropriate. various->mkToggle: updated to makeMapkeeper, along with related paradigm response. controllers->various: added gate use. controllers->various->various: added gateway restrictions, as appropriate. routes: updated mkToggle ref. models->fief->ruler: removed. various->npc: added park_id. models->npc->accessors: added Persona accessors. models->npc->homeTerritory: updated to home. models->park->getMapkeeperAttribute: updated to like, work. models->park->getMonarchAttribute: added. models->persona->action: updated to defaultAction. providers->Auth: added admin, mapkeeper, mapkeeperOwn, and own permission scopes. models->user->persona: added relationship. views->various->datatables: added permission wrap to appropriate actions. views->layouts->menu: removed unnecessary links. views->layouts->menu: categorized remaining links. views->layouts->menu: updated link icons.

+++ interQuest/app/Http/Controllers/API/CommentAPIController.php
@@ -11,6 +11,7 @@ use App\Http\Controllers\AppBaseController;
 use InfyOm\Generator\Criteria\LimitOffsetCriteria;
 use Prettus\Repository\Criteria\RequestCriteria;
 use Response;
+use Gate;
 
 /**
  * Class CommentController
@@ -99,6 +100,9 @@ class CommentAPIController extends AppBaseController
         if (empty($comment)) {
             return $this->sendError('Comment not found');
         }
+        if(Gate::denies('own', $comment->author_persona_id)){
+            return $this->sendError('Permission Denied');
+        }
 
         $comment = $this->commentRepository->update($input, $id);
 
@@ -120,6 +124,9 @@ class CommentAPIController extends AppBaseController
 
         if (empty($comment)) {
             return $this->sendError('Comment not found');
+        }
+        if(Gate::denies('own', $comment->author_persona_id)){
+            return $this->sendError('Permission Denied');
         }
 
         $comment->delete();
